refactor(core): extract maintenance/memory-role-report.ts — getMemoryRoleReport + compareMemoryRoleReports
